Refactor: Restructure project directory layout
- Move code from main.py to appropriate modules
- Create utils/ directory for utility functions
- Separate data processing logic into data_processing.py
- Update imports to reflect new file structure
