+++ .gitignore
@@ -76,3 +76,5 @@ $RECYCLE.BIN/
 *.lnk
 
 # End of https://www.toptal.com/developers/gitignore/api/windows,macos,linux
+/config/__pycache__
+/utils/__pycache__
